+++ tests/fixtures/probes.py
@@ -289,6 +289,21 @@ ProcEchoArrayProbe = {
     }
 }
 
+ProcEchoStrProbe = {
+    "type": "probe",
+    "name": "This probe is a process probe that simply echoes its arguments passed as a string",
+    "pauses": {
+        "before": 0,
+        "after": 0.1
+    },
+    "provider": {
+        "type": "process",
+        "path": sys.executable,
+        "arguments": "-c 'import sys; print(sys.argv)' --empty --number 1 --string 'with spaces' --string 'a second string with the same option'",
+        "timeout": 1
+    }
+}
+
 HTTPProbe = {
     "type": "probe",
     "name": "This probe is a HTTP probe",